Clean up for better Git Handling and fixed Mod Icons
Packed the 3000+ sprites into a ZIP file, please unpack outside of your
client directory to prevent readding it inadvertently.

Also attempted to rework the mod icons. Mod and Admin work, cannot get
the Owner to work. Checked for the source of where the handler for that
feature comes from (Class30_Sub2_Sub1_Sub1.java, lines 507-510) for
anyone willing to figure it out.

Also renamed the client launcher bat file to not be ".bat.bat" and added
!! to the beginning to bring to the top to not have to scroll through a
mile of classes and .java files.